blacklist_monitor/requirements.txt
Flask
APScheduler
dnspython
requests
